Writing StoreFindToken and BlobReadOptions in new formats
This commit writes StoreFindToken and BlobReadOptions in the new formats and thus
enables log segmentation end to end
