+++ ambry-store/src/main/java/com.github.ambry.store/PersistentIndex.java
@@ -17,11 +17,9 @@ import com.codahale.metrics.Timer;
 import com.github.ambry.config.StoreConfig;
 import com.github.ambry.utils.Time;
 import com.github.ambry.utils.Utils;
-import java.io.DataInputStream;
 import java.io.File;
 import java.io.FilenameFilter;
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Comparator;
@@ -1092,227 +1090,3 @@ class PersistentIndex {
   }
 }
 
-/**
- * The StoreFindToken is an implementation of FindToken.
- * It is used to provide a token to the client to resume
- * the find from where it was left previously. The StoreFindToken
- * maintains a offset to track entries within the journal. If the
- * offset gets outside the range of the journal, the storekey and
- * indexstartoffset that refers to the segment of the index is used
- * to perform the search. This is possible because the journal is
- * always equal or larger than the writable segment.
- */
-class StoreFindToken implements FindToken {
-  /**
-   * The type of the store token.
-   */
-  enum Type {
-    Uninitialized, JournalBased, IndexBased
-  }
-
-  private Type type;
-  private Offset offset;
-  private StoreKey storeKey;
-  private UUID sessionId;
-  private long bytesRead;
-
-  private static final short version = 0;
-  private static final int Version_Size = 2;
-  private static final int SessionId_Size = 4;
-  private static final int Offset_Size = 8;
-
-  private static final int Uninitialized_Offset = -1;
-
-  StoreFindToken() {
-    this(Type.Uninitialized, null, null, null);
-  }
-
-  StoreFindToken(StoreKey key, Offset indexSegmentStartOffset, UUID sessionId) {
-    this(Type.IndexBased, indexSegmentStartOffset, key, sessionId);
-  }
-
-  StoreFindToken(Offset offset, UUID sessionId) {
-    this(Type.JournalBased, offset, null, sessionId);
-  }
-
-  private StoreFindToken(Type type, Offset offset, StoreKey key, UUID sessionId) {
-    if (!type.equals(Type.Uninitialized)) {
-      if (offset == null || sessionId == null) {
-        throw new IllegalArgumentException("Offset [" + offset + "] or SessionId [" + sessionId + "] cannot be null");
-      } else if (type.equals(Type.IndexBased) && key == null) {
-        throw new IllegalArgumentException("StoreKey cannot be null for an index based token");
-      }
-    }
-    this.type = type;
-    this.offset = offset;
-    this.storeKey = key;
-    this.sessionId = sessionId;
-    this.bytesRead = -1;
-  }
-
-  void setBytesRead(long bytesRead) {
-    this.bytesRead = bytesRead;
-  }
-
-  static StoreFindToken fromBytes(DataInputStream stream, StoreKeyFactory factory) throws IOException {
-    StoreFindToken storeFindToken;
-    // read version
-    short version = stream.readShort();
-    switch (version) {
-      case 0:
-        // backwards compatibility
-        String logSegmentName = LogSegmentNameHelper.generateFirstSegmentName(1);
-        // read sessionId
-        String sessionId = Utils.readIntString(stream);
-        UUID sessionIdUUID = null;
-        if (!sessionId.isEmpty()) {
-          sessionIdUUID = UUID.fromString(sessionId);
-        }
-        // read offset
-        long offset = stream.readLong();
-        // read index start offset
-        long indexStartOffset = stream.readLong();
-        if (indexStartOffset != Uninitialized_Offset) {
-          // read store key if needed
-          storeFindToken = new StoreFindToken(factory.getStoreKey(stream), new Offset(logSegmentName, indexStartOffset),
-              sessionIdUUID);
-        } else if (offset != Uninitialized_Offset) {
-          storeFindToken = new StoreFindToken(new Offset(logSegmentName, offset), sessionIdUUID);
-        } else {
-          storeFindToken = new StoreFindToken();
-        }
-        break;
-      case 1:
-        // read sessionId
-        sessionId = Utils.readIntString(stream);
-        sessionIdUUID = null;
-        if (!sessionId.isEmpty()) {
-          sessionIdUUID = UUID.fromString(sessionId);
-        }
-        // read type
-        Type type = Type.values()[stream.readShort()];
-        switch (type) {
-          case Uninitialized:
-            storeFindToken = new StoreFindToken();
-            break;
-          case JournalBased:
-            Offset logOffset = Offset.fromBytes(stream);
-            storeFindToken = new StoreFindToken(logOffset, sessionIdUUID);
-            break;
-          case IndexBased:
-            Offset indexSegmentStartOffset = Offset.fromBytes(stream);
-            storeFindToken = new StoreFindToken(factory.getStoreKey(stream), indexSegmentStartOffset, sessionIdUUID);
-            break;
-          default:
-            throw new IllegalStateException("Unknown store find token type: " + type);
-        }
-        break;
-      default:
-        throw new IllegalArgumentException("Unrecognized version in StoreFindToken: " + version);
-    }
-    return storeFindToken;
-  }
-
-  Type getType() {
-    return type;
-  }
-
-  StoreKey getStoreKey() {
-    return storeKey;
-  }
-
-  Offset getOffset() {
-    return offset;
-  }
-
-  UUID getSessionId() {
-    return sessionId;
-  }
-
-  @Override
-  public long getBytesRead() {
-    if (this.bytesRead == -1) {
-      throw new IllegalStateException("Bytes read not initialized");
-    }
-    return this.bytesRead;
-  }
-
-  @Override
-  public byte[] toBytes() {
-    int size =
-        Version_Size + SessionId_Size + (sessionId == null ? 0 : sessionId.toString().getBytes().length) + Offset_Size
-            + Offset_Size + (storeKey == null ? 0 : storeKey.sizeInBytes());
-    byte[] buf = new byte[size];
-    ByteBuffer bufWrap = ByteBuffer.wrap(buf);
-    // add version
-    bufWrap.putShort(version);
-    // add sessionId
-    byte[] sessionIdBytes = sessionId == null ? null : sessionId.toString().getBytes();
-    bufWrap.putInt(sessionIdBytes == null ? 0 : sessionIdBytes.length);
-    if (sessionIdBytes != null) {
-      bufWrap.put(sessionIdBytes);
-    }
-    long logOffset = Uninitialized_Offset;
-    long indexStartOffset = Uninitialized_Offset;
-    if (type.equals(Type.JournalBased)) {
-      logOffset = offset.getOffset();
-    } else if (type.equals(Type.IndexBased)) {
-      indexStartOffset = offset.getOffset();
-    }
-    // add offset
-    bufWrap.putLong(logOffset);
-    // add index start offset
-    bufWrap.putLong(indexStartOffset);
-    // add storekey
-    if (storeKey != null) {
-      bufWrap.put(storeKey.toBytes());
-    }
-    return buf;
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("version: ").append(version);
-    sb.append("type: ").append(type);
-    if (!type.equals(Type.Uninitialized)) {
-      if (sessionId != null) {
-        sb.append(" sessionId ").append(sessionId);
-      }
-      if (storeKey != null) {
-        sb.append(" storeKey ").append(storeKey);
-      }
-      sb.append(" offset ").append(offset);
-      sb.append(" bytesRead ").append(bytesRead);
-    }
-    return sb.toString();
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-
-    StoreFindToken that = (StoreFindToken) o;
-
-    if (type != that.type) {
-      return false;
-    }
-    if (offset != null ? !offset.equals(that.offset) : that.offset != null) {
-      return false;
-    }
-    return storeKey != null ? storeKey.equals(that.storeKey) : that.storeKey == null;
-  }
-
-  @Override
-  public int hashCode() {
-    int result = type.hashCode();
-    result = 31 * result + (offset != null ? offset.hashCode() : 0);
-    result = 31 * result + (storeKey != null ? storeKey.hashCode() : 0);
-    return result;
-  }
-}
