+++ ambry-store/src/main/java/com.github.ambry.store/StoreMessageReadSet.java
@@ -41,9 +41,8 @@ class BlobReadOptions implements Comparable<BlobReadOptions>, Closeable {
   private final AtomicBoolean open = new AtomicBoolean(true);
   private final Logger logger = LoggerFactory.getLogger(getClass());
 
-  private static final short VERSION = 0;
+  private static final short VERSION = 1;
   private static final short VERSION_LENGTH = 2;
-  private static final short OFFSET_LENGTH = 8;
   private static final short SIZE_LENGTH = 8;
   private static final short EXPIRES_AT_MS_LENGTH = 8;
 
@@ -100,10 +99,12 @@ class BlobReadOptions implements Comparable<BlobReadOptions>, Closeable {
   }
 
   byte[] toBytes() {
-    byte[] buf = new byte[VERSION_LENGTH + OFFSET_LENGTH + SIZE_LENGTH + EXPIRES_AT_MS_LENGTH + storeKey.sizeInBytes()];
+    byte[] offsetBytes = offset.toBytes();
+    byte[] buf =
+        new byte[VERSION_LENGTH + offsetBytes.length + SIZE_LENGTH + EXPIRES_AT_MS_LENGTH + storeKey.sizeInBytes()];
     ByteBuffer bufWrap = ByteBuffer.wrap(buf);
     bufWrap.putShort(VERSION);
-    bufWrap.putLong(offset.getOffset());
+    bufWrap.put(offsetBytes);
     bufWrap.putLong(size);
     bufWrap.putLong(expiresAtMs);
     bufWrap.put(storeKey.toBytes());
